Backend/requirements.txt
bcrypt==3.2.0
python-dotenv==0.20.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.3.0
PyJWT==2.3.0
pymongo[srv]==4.4.1
requests==2.27.1
pydantic==1.9.0
pytest==7.0.1
Jinja2==3.0.3
Werkzeug==2.0.2
openai==0.27.0


# activating or deactivating virtual env
# myenv\Scripts\activate
# deactivate
